Fix MiniMax H3 parity and generic INT8 scales
Match ComfyUI float32 timestep and velocity conversion semantics. Accept strict scalar or rowwise non-ConvRot INT8 scales while preserving ConvRot rowwise requirements. Remove temporary H3 tracing hooks and synchronize receipt tests with the production defaults.

Reviewed: API cohesion, quantization invariants, sampling-engine layering, diagnostic removal, and documentation impact; no findings.
Amp-Thread-ID: https://ampcode.com/threads/T-01a0c878-a553-76bb-8728-993eae165094

+++ packages/dinkster-inference-torch/src/dinkster_inference_torch/minimax_h3_runtime.py
@@ -1710,7 +1710,7 @@ class MiniMaxH3DiTRuntime(MultiStreamSamplingRuntime):
                             attention_kernel_factory=attention_kernel_factory,
                         )
                 packed_velocity, _ = pack_latent_streams(velocity)
-                result = x - packed_velocity * sigma
+                result = x - packed_velocity.float() * sigma
             _check_cancelled(cancelled)
             return result
